requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
arcp==0.2.1
argo-workflows==6.5.8
attrs==23.2.0
beautifulsoup4==4.12.3
bioblend==1.3.0
CacheControl==0.14.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
CordraPy @ git+https://github.com/usnistgov/CordraPy@7d73f3f58461a2dc7540970fe7defaa361b61740
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
frozenlist==1.4.1
galaxy2cwl==0.1.4
gxformat2==0.18.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
orjson==3.10.6
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
PyYAML==6.0.1
rdflib==7.0.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rocrate==0.10.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
schema-salad==8.7.20240718183047
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tinydb==4.8.0
tuspy==1.0.3
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
